5072 - Rename latest activities to recent highlights
