--- src/client/pages/Kiosk/LatestActivities/LatestActivities.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-import MapWrapper from 'client/pages/Kiosk/LatestActivities/MapWrapper'
-
-const LatestActivities: React.FC = () => {
-  return <MapWrapper />
-}
-
-export default LatestActivities
