--- src/client/pages/Kiosk/LatestActivities/index.ts
@@ -1,1 +1,0 @@
-export { default } from './LatestActivities'
