+++ src/client/pages/Kiosk/RecentHighlights/ActivityList/ActivityList.scss
