+++ src/client/pages/Kiosk/RecentHighlights/ActivityList/index.ts
