+++ src/client/pages/Kiosk/RecentHighlights/ActivityListItem/ActivityListItem.scss
