+++ src/client/pages/Kiosk/RecentHighlights/ActivityListItem/index.ts
