+++ src/client/pages/Kiosk/RecentHighlights/Map/Map.scss
