+++ src/client/pages/Kiosk/RecentHighlights/Map/hooks/useLatestActivitiesMap.ts
