+++ src/client/pages/Kiosk/RecentHighlights/Map/index.ts
