+++ src/client/pages/Kiosk/RecentHighlights/MapWrapper/MapWrapper.scss
