+++ src/client/pages/Kiosk/RecentHighlights/MapWrapper/index.ts
