+++ src/client/pages/Kiosk/RecentHighlights/RecentHighlights.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import MapWrapper from 'client/pages/Kiosk/RecentHighlights/MapWrapper'
+
+const RecentHighlights: React.FC = () => {
+  return <MapWrapper />
+}
+
+export default RecentHighlights
