+++ src/client/pages/Kiosk/RecentHighlights/RecentHighlights/RecentHighlights.scss
