+++ src/client/pages/Kiosk/RecentHighlights/RecentHighlights/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RecentHighlights'
