+++ src/client/pages/Kiosk/RecentHighlights/index.ts
